NODE-2: address KeystoreOperation error
The instructions were previously using a local directory under $HOME to
save various keys. This was causing an issue at the time of generating
them because the user inside the container did not have write access to
that path.

A common workaround for this is to `chmod 777` the directory, but that
would leave the keys open to any other  user on the system. Instead,
we'll use docker volumes to store the keys.